requirements.txt
mss~=6.1.0
Pillow~=9.2.0
pynput~=1.7.6